Split up files and implement and optimize GitLabScanner
