+++ scanners/git-repo-scanner/scanner/git_repo_scanner/abstract_scanner.py
@@ -1,0 +1,38 @@
+import abc
+from datetime import datetime
+from typing import Dict, List, Optional
+
+FINDING = Dict[str, any]
+
+
+class AbstractScanner(abc.ABC):
+
+    @property
+    @abc.abstractmethod
+    def git_type(self) -> str:
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def process(self, start_time: Optional[datetime] = None, end_time: Optional[datetime] = None) -> List[FINDING]:
+        raise NotImplementedError()
+
+    def _create_finding(self, repo_id: str, web_url: str, full_name: str, owner_type: str, owner_id: str,
+                        owner_name: str, created_at: str, last_activity_at: str, visibility: str) -> FINDING:
+        return {
+            'name': f'{self.git_type} Repo',
+            'description': f'A {self.git_type} repository',
+            'category': 'Git Repository',
+            'osi_layer': 'APPLICATION',
+            'severity': 'INFORMATIONAL',
+            'attributes': {
+                'id': repo_id,
+                'web_url': web_url,
+                'full_name': full_name,
+                'owner_type': owner_type,
+                'owner_id': owner_id,
+                'owner_name': owner_name,
+                'created_at': created_at,
+                'last_activity_at': last_activity_at,
+                'visibility': visibility
+            }
+        }
